handlers/test.py
import os, google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()
genai.configure(api_key=os.getenv("GEMINI_API_KEY"))

model = genai.GenerativeModel("gemini-1.5-flash")
resp = model.generate_content("Hello Gemini, are you working?")
print(resp.text)
